--- dev/i18n.ts
@@ -1,16 +1,0 @@
-import translations from "./language/translations.json";
-import { createGettext } from "/@gettext/";
-
-const gettext = createGettext({
-    availableLanguages: {
-        en_GB: "British English",
-        fr_FR: "Français",
-        it_IT: "Italiano",
-        zh_CN: "简体中文",
-    },
-    defaultLanguage: "en_GB",
-    translations: translations,
-    setGlobalProperties: true,
-});
-
-export { gettext };
